requirements.txt
streamlit==1.41.1
langchain==0.3.14
langchain-community==0.3.14
langchain-text-splitters==0.3.4
transformers==4.48.2
accelerate>=0.26.0
sentence-transformers==3.3.1
torch>=2.1.0
faiss-cpu==1.8.0.post1; python_version < "3.13"
faiss-cpu>=1.10.0; python_version >= "3.13"
gitpython==3.1.44
